refactor: Serializable을 구현하는 클래스를 Parcelable을 구현하도록 변경
